+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Allow transformers to be applied to namers
 * Add Const and Replace transformers
 * Add `useHealthCheck` option to marathon namer
+* Add experimental Apache Curator namer
 
 ## 0.8.2